UF-9159: Add municipality ID.

+++ src/main/java/se/sundsvall/contactsettings/service/DelegateService.java
@@ -1,6 +1,5 @@
 package se.sundsvall.contactsettings.service;
 
-import static java.lang.String.format;
 import static java.util.Collections.emptyList;
 import static java.util.Objects.isNull;
 import static java.util.Objects.nonNull;
@@ -13,8 +12,8 @@ import static se.sundsvall.contactsettings.service.Constants.ERROR_MESSAGE_DELEG
 import static se.sundsvall.contactsettings.service.Constants.ERROR_MESSAGE_PRINCIPAL_NOT_FOUND;
 import static se.sundsvall.contactsettings.service.mapper.DelegateMapper.toDelegate;
 import static se.sundsvall.contactsettings.service.mapper.DelegateMapper.toDelegateEntity;
+import static se.sundsvall.contactsettings.service.mapper.DelegateMapper.toDelegateList;
 
-import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.stereotype.Service;
@@ -25,8 +24,6 @@ import se.sundsvall.contactsettings.api.model.DelegateCreateRequest;
 import se.sundsvall.contactsettings.api.model.FindDelegatesParameters;
 import se.sundsvall.contactsettings.integration.db.ContactSettingRepository;
 import se.sundsvall.contactsettings.integration.db.DelegateRepository;
-import se.sundsvall.contactsettings.integration.db.model.DelegateEntity;
-import se.sundsvall.contactsettings.service.mapper.DelegateMapper;
 
 @Service
 public class DelegateService {
@@ -39,64 +36,73 @@ public class DelegateService {
 		this.contactSettingRepository = contactSettingRepository;
 	}
 
-	public Delegate create(final DelegateCreateRequest delegateCreateRequest) {
+	public Delegate create(final String municipalityId, final DelegateCreateRequest delegateCreateRequest) {
 
 		// Verifications:
-		verifyThatAgentExists(delegateCreateRequest.getAgentId());
-		verifyThatPrincipalExists(delegateCreateRequest.getPrincipalId());
+		verifyThatAgentExists(municipalityId, delegateCreateRequest.getAgentId());
+		verifyThatPrincipalExists(municipalityId, delegateCreateRequest.getPrincipalId());
 		verifyThatDelegateDoesNotAlreadyExist(delegateCreateRequest.getPrincipalId(), delegateCreateRequest.getAgentId());
 
 		// All good: proceed
 		return toDelegate(delegateRepository.save(toDelegateEntity(delegateCreateRequest)));
 	}
 
-	public Delegate read(final String id) {
+	public Delegate read(final String municipalityId, final String id) {
 
 		// Fetch/validate
-		final var delegateEntity = delegateRepository.findById(id).orElseThrow(() -> Problem.valueOf(NOT_FOUND, format(ERROR_MESSAGE_DELEGATE_NOT_FOUND, id)));
+		final var entity = delegateRepository.findById(id).orElseThrow(() -> Problem.valueOf(NOT_FOUND, ERROR_MESSAGE_DELEGATE_NOT_FOUND.formatted(id)));
+
+		if (!entity.getAgent().getMunicipalityId().equals(municipalityId) || !entity.getPrincipal().getMunicipalityId().equals(municipalityId)) {
+			throw Problem.valueOf(NOT_FOUND, ERROR_MESSAGE_DELEGATE_NOT_FOUND.formatted(id));
+		}
 
 		// All good: proceed
-		return toDelegate(delegateEntity);
+		return toDelegate(entity);
 	}
 
-	public void delete(final String id) {
+	public void delete(final String municipalityId, final String id) {
 
 		// Fetch/validate
-		final var entity = delegateRepository.findById(id).orElseThrow(() -> Problem.valueOf(NOT_FOUND, format(ERROR_MESSAGE_DELEGATE_NOT_FOUND, id)));
+		final var entity = delegateRepository.findById(id).orElseThrow(() -> Problem.valueOf(NOT_FOUND, ERROR_MESSAGE_DELEGATE_NOT_FOUND.formatted(id)));
+
+		if (!entity.getAgent().getMunicipalityId().equals(municipalityId) || !entity.getPrincipal().getMunicipalityId().equals(municipalityId)) {
+			throw Problem.valueOf(NOT_FOUND, ERROR_MESSAGE_DELEGATE_NOT_FOUND.formatted(id));
+		}
 
 		// All good: proceed
 		delegateRepository.delete(entity);
 	}
 
-	public List<Delegate> find(final FindDelegatesParameters parameters) {
+	public List<Delegate> find(final String municipalityId, final FindDelegatesParameters parameters) {
 		if (isNull(parameters)) {
 			return emptyList();
 		}
-
-		final var result = new ArrayList<DelegateEntity>();
 		if (allNotNull(parameters.getAgentId(), parameters.getPrincipalId())) {
-			result.addAll(delegateRepository.findByPrincipalIdAndAgentId(parameters.getPrincipalId(), parameters.getAgentId()));
-		} else if (nonNull(parameters.getAgentId())) {
-			result.addAll(delegateRepository.findByAgentId(parameters.getAgentId()));
-		} else if (nonNull(parameters.getPrincipalId())) {
-			result.addAll(delegateRepository.findByPrincipalId(parameters.getPrincipalId()));
+			verifyThatAgentExists(municipalityId, parameters.getAgentId());
+			verifyThatPrincipalExists(municipalityId, parameters.getPrincipalId());
+			return toDelegateList(delegateRepository.findByPrincipalIdAndAgentId(parameters.getPrincipalId(), parameters.getAgentId()));
+		}
+		if (nonNull(parameters.getAgentId())) {
+			verifyThatAgentExists(municipalityId, parameters.getAgentId());
+			return toDelegateList(delegateRepository.findByAgentId(parameters.getAgentId()));
+		}
+		if (nonNull(parameters.getPrincipalId())) {
+			verifyThatPrincipalExists(municipalityId, parameters.getPrincipalId());
+			return toDelegateList(delegateRepository.findByPrincipalId(parameters.getPrincipalId()));
 		}
 
-		return result.stream()
-			.distinct()
-			.map(DelegateMapper::toDelegate)
-			.toList();
+		return emptyList();
 	}
 
-	private void verifyThatAgentExists(String agentId) {
-		if (!contactSettingRepository.existsById(agentId)) {
-			throw Problem.valueOf(NOT_FOUND, format(ERROR_MESSAGE_AGENT_NOT_FOUND, agentId));
+	private void verifyThatAgentExists(String municipalityId, String agentId) {
+		if (!contactSettingRepository.existsByMunicipalityIdAndId(municipalityId, agentId)) {
+			throw Problem.valueOf(NOT_FOUND, ERROR_MESSAGE_AGENT_NOT_FOUND.formatted(agentId));
 		}
 	}
 
-	private void verifyThatPrincipalExists(String principalId) {
-		if (!contactSettingRepository.existsById(principalId)) {
-			throw Problem.valueOf(NOT_FOUND, format(ERROR_MESSAGE_PRINCIPAL_NOT_FOUND, principalId));
+	private void verifyThatPrincipalExists(String municipalityId, String principalId) {
+		if (!contactSettingRepository.existsByMunicipalityIdAndId(municipalityId, principalId)) {
+			throw Problem.valueOf(NOT_FOUND, ERROR_MESSAGE_PRINCIPAL_NOT_FOUND.formatted(principalId));
 		}
 	}
